requirements.txt
wheel 
Pillow
django
djangorestframework
djangorestframework-simplejwt

drf-yasg
python-decouple
python-slugify

redis 
celery